Responded to a subset of PR review comments

+++ lib/gauge_loop_trace.cu
@@ -23,14 +23,12 @@ namespace quda {
       factor(factor),
       p(p)
     {
-      if (p.num_paths != loop_traces.size()) errorQuda("Loop traces size %lu != number of paths %d", loop_traces.size(), p.num_paths);
+      if (p.num_paths != static_cast<int>(loop_traces.size())) errorQuda("Loop traces size %lu != number of paths %d", loop_traces.size(), p.num_paths);
 
       strcat(aux, "num_paths=");
       char loop_str[4];
       u32toa(loop_str, p.num_paths);
       strcat(aux, loop_str);
-
-      // we should add some hash of lengths, paths, etc...
 
       apply(device::get_default_stream());
     }
